# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dlavie Chat API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/{messageId}:
    patch:
      operationId: editMessage
      tags: [messages]
      summary: Edit a message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required: [content]
      responses:
        "200":
          description: Edited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /messages/{messageId}/read:
    post:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark message as read
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /messages/{messageId}/reactions:
    post:
      operationId: addReaction
      tags: [messages]
      summary: Add reaction to message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emoji:
                  type: string
              required: [emoji]
      responses:
        "200":
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removeReaction
      tags: [messages]
      summary: Remove reaction from message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emoji:
                  type: string
              required: [emoji]
      responses:
        "200":
          description: Reaction removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # POSTS
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                  unreadCount:
                    type: integer
                required: [notifications, unreadCount]
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # UPLOAD
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    UserPublic:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
          nullable: true
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isOnline:
          type: boolean
        lastSeenAt:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, moderator, admin]
      required: [id, displayName, isOnline, role]
...
    MessagePreview:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
          nullable: true
        type:
          type: string
        senderName:
          type: string
        createdAt:
          type: string
      required: [id, type, senderName, createdAt]
...
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        type:
          type: string
          enum: [text, image, video, file, audio, voice, system]
        content:
          type: string
          nullable: true
        mediaUrl:
          type: string
          nullable: true
        replyToMessageId:
          type: string
          nullable: true
        replyTo:
          $ref: "#/components/schemas/MessagePreview"
          nullable: true
        editedAt:
          type: string
          nullable: true
        deletedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        sender:
          $ref: "#/components/schemas/UserPublic"
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        status:
          type: string
          enum: [sent, delivered, read]
          nullable: true
      required: [id, conversationId, senderId, type, createdAt, sender, reactions]
...
    Reaction:
      type: object
      properties:
        emoji:
          type: string
        count:
          type: integer
        users:
          type: array
          items:
            type: string
      required: [emoji, count, users]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
        relatedType:
          type: string
          nullable: true
        relatedId:
          type: string
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, type, title, body, isRead, createdAt]
...